requirements.txt
PyYAML>=6.0
fastapi>=0.110.0
pydantic>=2.6.0
uvicorn>=0.27.0
numpy>=1.26.0
